+++ tests/RecurringGeneratorsProviderTest.php
@@ -85,4 +85,62 @@ class RecurringGeneratorsProviderTest extends \Fakturoid\Tests\TestCase
         $response = $provider->create(['page' => 2]);
         $this->assertEquals(['page' => 2], $response->getBody(true));
     }
+
+    public function testPause(): void
+    {
+        $dispatcher = $this->createMock(Dispatcher::class);
+
+        $id = 6;
+        $responseInterface = $this->createPsrResponseMock(200, 'application/json', '{"active": false}');
+        $dispatcher->expects($this->once())
+            ->method('patch')
+            ->with(sprintf('/accounts/{accountSlug}/recurring_generators/%d/pause.json', $id))
+            ->willReturn(new Response($responseInterface));
+
+        $provider = new RecurringGeneratorsProvider($dispatcher);
+        $response = $provider->pause($id);
+        $this->assertEquals(['active' => false], $response->getBody(true));
+    }
+
+    public function testActivate(): void
+    {
+        $dispatcher = $this->createMock(Dispatcher::class);
+
+        $id = 6;
+        $responseInterface = $this->createPsrResponseMock(200, 'application/json', '{"active": true}');
+        $dispatcher->expects($this->once())
+            ->method('patch')
+            ->with(sprintf('/accounts/{accountSlug}/recurring_generators/%d/activate.json', $id))
+            ->willReturn(new Response($responseInterface));
+
+        $provider = new RecurringGeneratorsProvider($dispatcher);
+        $response = $provider->activate($id);
+        $this->assertEquals(['active' => true], $response->getBody(true));
+    }
+
+    public function testActivateWithNextOccurrenceDate(): void
+    {
+        $dispatcher = $this->createMock(Dispatcher::class);
+
+        $id = 6;
+        $responseInterface = $this->createPsrResponseMock(
+            200,
+            'application/json',
+            '{"active": true, "next_occurrence_on": "2025-02-15"}'
+        );
+        $dispatcher->expects($this->once())
+            ->method('patch')
+            ->with(
+                sprintf('/accounts/{accountSlug}/recurring_generators/%d/activate.json', $id),
+                ['next_occurrence_on' => '2025-02-15']
+            )
+            ->willReturn(new Response($responseInterface));
+
+        $provider = new RecurringGeneratorsProvider($dispatcher);
+        $response = $provider->activate($id, ['next_occurrence_on' => '2025-02-15']);
+        $this->assertEquals(
+            ['active' => true, 'next_occurrence_on' => '2025-02-15'],
+            $response->getBody(true)
+        );
+    }
 }